Moved more server start logic out of extension.ts.
